fix: 10 runtime bugs + modularize the largest files (coverage 97%→98%)
Bug fixes (verified against a live install; no unique_id changes):
- Motion never reported in direct (SQS) mode — handler wrote `motion` but the
  sensor reads `motion_detected` (the key the SSE handler already writes).
- Diagnostics download crashed for every proxy/SSE user — `_connectivity_snapshot`
  tried to *call* the `is_connected` @property (a bool).
- CombiProtect glass-break was silent in SQS mode (only read the SSE key).
- Door-sensor fast-poll task could die on concurrent hub discovery
  (dictionary-changed-size race); the loop now snapshots the spaces.
- DoorProtect Plus / GlassProtect config switches bounced back — the poller
  overwrote 5 optimistic attributes without checking the optimistic guard.
- Doorbell "Last ring" TIMESTAMP sensor went unknown on every ring (ISO string
  vs datetime); the value is now parsed back to a datetime.
- Duplicate hub_battery unique_id collision — one entity was silently dropped.
- Manual Call Point colour ENUM was missing yellow/green/graphite.
- Socket external-power was wired to a never-populated key; SSE wrote a third,
  dead key. All now agree on `externally_powered`.
- Auth failures on the rooms endpoint bypassed the reauth threshold.

Refactor (behaviour-preserving, unique_id stable):
- Event tables -> event_maps.py (single source of truth for SSE + SQS).
- Door fast-poll -> _coordinator_door_poll.py (AjaxDoorPollingMixin).
- Device normalisation -> _device_normalize.py (AjaxDeviceNormalizeMixin).
- Switch entities -> _switch_entity.py / _switch_dimmer.py.
- Options flow -> config_flow_options.py.
- Service registration -> _services.py.

Tests 1826 -> 1833, coverage 97% -> 98% (__init__ and coordinator to ~100%).
ruff, ruff-format, mypy --strict and hassfest all clean.
